using Integer data type for nullable columns in Event to indicate null as no value instead of zero

+++ opennms-webapp/src/main/java/org/opennms/web/event/Event.java
@@ -482,8 +482,8 @@ public class Event {
      *
      * @return a int.
      */
-    public int getNodeId() {
-        return (nodeID.intValue());
+    public Integer getNodeId() {
+        return (nodeID);
     }
 
     /**
@@ -500,8 +500,8 @@ public class Event {
      *
      * @return a int.
      */
-    public int getServiceId() {
-        return (serviceID.intValue());
+    public Integer getServiceId() {
+        return (serviceID);
     }
 
     /**
